Move endpointConfig to configure
